Move footer + navbar into layout directory
